feat: implemented creation of ticket

+++ LAB03 - Ticket management/server/src/main/kotlin/com/lab3/server/controller/ProductController.kt
@@ -22,7 +22,7 @@ class ProductController @Autowired constructor(val productService: ProductServic
 
     @GetMapping("/API/products/{productId}")
     @ResponseStatus(HttpStatus.OK)
-    fun getProductById(@PathVariable("productId") productId: Int): ProductDTO? {
+    fun getProductById(@PathVariable("productId") productId: Long): ProductDTO? {
         return productService.getProductById(productId)
             ?: throw Exception.ProductNotFoundException("No product matched the requested Id")
     }
