+++ LAB03 - Ticket management/server/src/main/kotlin/com/lab3/server/repository/CustomerRepository.kt
@@ -3,6 +3,6 @@ package com.lab3.server.repository
 import com.lab3.server.model.Customer
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface CustomerRepository : JpaRepository<Customer, Int> {
+interface CustomerRepository : JpaRepository<Customer, Long> {
     fun findByEmail(email: String): Customer?
 }
